backend/services/experiment-service/pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "experiment-service"
version = "0.1.0"
description = "Experiment Tracking backend service aligned with Experiment Tracking Platform spec"
authors = ["Product Dev Course"]
readme = "README.md"
packages = [
    { include = "experiment_service", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.10.5"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
orjson = "^3.10.3"
structlog = "^24.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.21.1"
pytest-aiohttp = "^1.0.5"
httpx = "^0.27.0"
ruff = "^0.5.0"
mypy = "^1.10.0"
yandex-taxi-testsuite = {extras = ["postgresql"], version = "^0.3.9"}
openapi-generator-cli = "^7.17.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-p testsuite.pytest_plugin"
testpaths = ["tests"]
pythonpath = ["src"]


